Add: `formatColonyRoles` helper
